V.0.0.1.43 [PRE-ALPHA] Diamond2DGL on process: add buffer resize/rescale correctly, add Black Diamonds game demo

+++ src/main/java/diamond2DGL/Camera.java
@@ -8,8 +8,8 @@ public class Camera {
 
     // ATTRIBUTES
     private Matrix4f pMatrix, vMatrix, invProj, invView;
-    private float pWidth = 6f;
-    private float pHeight = 3f;
+    private int pWidth = 6;
+    private int pHeight = 3;
     private Vector2f projectionSize = new Vector2f(pWidth, pHeight);
     private float zoom = 1.0f;
     public Vector2f pos;
